// src/enums/response-code.enum.ts
/**
 * Format:
 *  + Key: ModuleName_ResponseName
 *  + Value: Prefix + {000->999}
 * Prefix:
 *  + 01: Common
 *  + 02: Auth Module
 *  + 03: User Module
 */
export enum RESPONSE_CODE {
  // Common
  COMMON_CORS_NOT_ALLOW = '01001',
  COMMON_VALIDATION_FAILED = '01002',
  COMMON_BAD_REQUEST = '01003',
  COMMON_NOT_FOUND = '01004',
  COMMON_ERROR = '01005',

  // Auth Module
  AUTH_TOKEN_MISSING = '02001',
  AUTH_INVALID_TOKEN = '02002',
  AUTH_INVALID_USERNAME_PASSWORD = '02003',
  AUTH_INVALID_REFRESH_TOKEN = '02004',
  AUTH_LOGOUT_SUCCESS = '02005',
  AUTH_EMAIL_MISSING = '02006',
  AUTH_INVALID_EMAIL = '02007',
  AUTH_PASSWORD_MISSING = '02008',
  AUTH_REFRESH_TOKEN_MISSING = '02009',

  // User Module
  USER_EMAIL_EXISTED = '03001',
  USER_PASSWORD_AT_LEAST_CHAR = '03002',
  USER_PASSWORD_CONTAIN_UPPER_CHAR = '03003',
  USER_PASSWORD_CONTAIN_NUMBER = '03004',
  USER_EMAIL_MISSING = '03005',
  USER_INVALID_EMAIL = '03006',
  USER_FIRSTNAME_MISSING = '03007',
  USER_FIRSTNAME_STRING = '03008',
  USER_FIRSTNAME_CHAR_LENGTH = '03009',
}